Add delete endpoints for Income and Expense, logs of all operations

+++ backend/backend-app/src/main/java/ru/rgasymov/moneymanager/service/impl/ExpenseTypeService.java
@@ -1,6 +1,7 @@
 package ru.rgasymov.moneymanager.service.impl;
 
 import java.util.Set;
+import javax.validation.ValidationException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -10,6 +11,7 @@ import ru.rgasymov.moneymanager.domain.dto.response.ExpenseTypeResponseDto;
 import ru.rgasymov.moneymanager.domain.entity.ExpenseType;
 import ru.rgasymov.moneymanager.domain.entity.User;
 import ru.rgasymov.moneymanager.mapper.ExpenseTypeMapper;
+import ru.rgasymov.moneymanager.repository.ExpenseRepository;
 import ru.rgasymov.moneymanager.repository.ExpenseTypeRepository;
 import ru.rgasymov.moneymanager.service.TypeService;
 import ru.rgasymov.moneymanager.service.UserService;
@@ -21,9 +23,19 @@ public class ExpenseTypeService implements TypeService<ExpenseTypeRequestDto, Ex
 
     private final ExpenseTypeRepository expenseTypeRepository;
 
+    private final ExpenseRepository expenseRepository;
+
     private final UserService userService;
 
     private final ExpenseTypeMapper expenseTypeMapper;
+
+    @Transactional(readOnly = true)
+    @Override
+    public Set<ExpenseTypeResponseDto> findAll() {
+        User currentUser = userService.getCurrentUser();
+        Set<ExpenseType> result = expenseTypeRepository.findAllByUserId(currentUser.getId());
+        return expenseTypeMapper.toDtos(result);
+    }
 
     @Transactional
     @Override
@@ -41,14 +53,10 @@ public class ExpenseTypeService implements TypeService<ExpenseTypeRequestDto, Ex
     @Override
     public void delete(Long id) {
         User currentUser = userService.getCurrentUser();
-        expenseTypeRepository.deleteByIdAndUserId(id, currentUser.getId());
-    }
 
-    @Transactional(readOnly = true)
-    @Override
-    public Set<ExpenseTypeResponseDto> findAll() {
-        User currentUser = userService.getCurrentUser();
-        Set<ExpenseType> result = expenseTypeRepository.findAllByUserId(currentUser.getId());
-        return expenseTypeMapper.toDtos(result);
+        if (expenseRepository.existsByExpenseTypeId(id)) {
+            throw new ValidationException("Could not delete an expense type while it is being referenced by any expenses");
+        }
+        expenseTypeRepository.deleteByIdAndUserId(id, currentUser.getId());
     }
 }
