+++ backend/backend-app/src/main/java/ru/rgasymov/moneymanager/service/impl/IncomeTypeService.java
@@ -1,6 +1,7 @@
 package ru.rgasymov.moneymanager.service.impl;
 
 import java.util.Set;
+import javax.validation.ValidationException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -10,6 +11,7 @@ import ru.rgasymov.moneymanager.domain.dto.response.IncomeTypeResponseDto;
 import ru.rgasymov.moneymanager.domain.entity.IncomeType;
 import ru.rgasymov.moneymanager.domain.entity.User;
 import ru.rgasymov.moneymanager.mapper.IncomeTypeMapper;
+import ru.rgasymov.moneymanager.repository.IncomeRepository;
 import ru.rgasymov.moneymanager.repository.IncomeTypeRepository;
 import ru.rgasymov.moneymanager.service.TypeService;
 import ru.rgasymov.moneymanager.service.UserService;
@@ -21,9 +23,19 @@ public class IncomeTypeService implements TypeService<IncomeTypeRequestDto, Inco
 
     private final IncomeTypeRepository incomeTypeRepository;
 
+    private final IncomeRepository incomeRepository;
+
     private final UserService userService;
 
     private final IncomeTypeMapper incomeTypeMapper;
+
+    @Transactional(readOnly = true)
+    @Override
+    public Set<IncomeTypeResponseDto> findAll() {
+        User currentUser = userService.getCurrentUser();
+        Set<IncomeType> result = incomeTypeRepository.findAllByUserId(currentUser.getId());
+        return incomeTypeMapper.toDtos(result);
+    }
 
     @Transactional
     @Override
@@ -41,14 +53,10 @@ public class IncomeTypeService implements TypeService<IncomeTypeRequestDto, Inco
     @Override
     public void delete(Long id) {
         User currentUser = userService.getCurrentUser();
-        incomeTypeRepository.deleteByIdAndUserId(id, currentUser.getId());
-    }
 
-    @Transactional(readOnly = true)
-    @Override
-    public Set<IncomeTypeResponseDto> findAll() {
-        User currentUser = userService.getCurrentUser();
-        Set<IncomeType> result = incomeTypeRepository.findAllByUserId(currentUser.getId());
-        return incomeTypeMapper.toDtos(result);
+        if (incomeRepository.existsByIncomeTypeId(id)) {
+            throw new ValidationException("Could not delete an income type while it is being referenced by any incomes");
+        }
+        incomeTypeRepository.deleteByIdAndUserId(id, currentUser.getId());
     }
 }
